Move detectBlockType to MediaBlockType enum, and extract enum
